minor: if a size_t is used with a method that uses int/uint and which cannot be changed, make the cast explicit

+++ src/data_tag.c
@@ -63,7 +63,7 @@ static void gpsd_client_event(struct mg_connection *nc, int ev, void *ev_data)
             // Success
             fprintf(stderr, "GPSd Connected...\n");
             if (ctx->init_str && *ctx->init_str) {
-                mg_send(nc, ctx->init_str, strlen(ctx->init_str));
+                mg_send(nc, ctx->init_str, (int)strlen(ctx->init_str));
             }
         }
         else {
