+++ src/devices/honeywell_cm921.c
@@ -54,7 +54,7 @@ static data_t *add_hex_string(data_t *data, const char *name, const uint8_t *buf
 {
     if (buf && buf_sz > 0) {
         char tstr[256];
-        bitrow_snprint(buf, buf_sz * 8, tstr, sizeof (tstr));
+        bitrow_snprint(buf, (unsigned int)buf_sz * 8, tstr, sizeof (tstr));
         data = data_append(data, name, "", DATA_STRING, tstr, NULL);
     }
     return data;
